Restructured foldering and naming conventions for webapp
