--- main.py
@@ -1,11 +1,0 @@
-import pickle
-
-def load_model():
-    return
-    # with open('model.pkl', 'rb') as file:
-    #     data = pickle.load(file)
-    # return data
-data = load_model()
-
-if __name__ == '__main__':
-    model = None
